A new option to filter the headings we want to feature on the table of content.
It may be useful if, for instance, we want to exclude headings
that are within elements such as <article> or <section>. The value must
be a function, whose first argument is the jquery collection of headings
and that return a sub-collection.

An example: remove all the headings that are within an article element.

    $("#toc").tableOfContents("#main-content", {
        startLevel: 2,
        depth: 1,
        exclude: function ($headings, $scope) {
            return $headings.not(function() { return $(this).parentsUntil($scope).is('article') })
        }
    });